Conflicts:
	include/write_dndt_ase.h
	src/main.cu
	src/write_dndt_ase.cu

+++ include/write_matlab_output.h
@@ -21,7 +21,7 @@
 void writeMatlabOutput(
     std::vector<float>* ase,
     std::vector<unsigned> *N_rays, 
-    std::vector<float> *expectedValues,
+    std::vector<double> *expectedValues,
     unsigned numberOfWavelengths,
     unsigned numberOfSamples);
